refactor: tables folder structure
