+++ src/index.ts
@@ -3,7 +3,7 @@
 export * from "@/createWrapper";
 
 // Local tables (with custom schema and templates)
-export * from "@/tables/createLocalTable";
+export * from "@/tables/core/createLocalTable";
 
 // Queries: global (filtering keys and callbacks), query (direct query returning Entity[]), useQuery (React hook returning Entity[] with callbacks)
 export { $query, query, useQuery } from "@/queries";
