+++ src/queries/index.ts
@@ -1,5 +1,3 @@
-export * from "@/queries/createTableWatcher";
-
 export * from "@/queries/query";
 export * from "@/queries/$query";
 export * from "@/queries/useQuery";
