+++ src/tables/core/createLocalTable.ts
