+++ src/tables/index.ts
@@ -1,6 +1,9 @@
-export * from "@/tables/createTable";
-export * from "@/tables/createLocalTable";
-export * from "@/tables/createContractTables";
-export * from "@/tables/createTableKeyMethods";
-export * from "@/tables/createTableMethods";
+export * from "@/tables/core/createTable";
+export * from "@/tables/core/createLocalTable";
+export * from "@/tables/core/createContractTables";
+
+export * from "@/tables/methods/createTableKeyMethods";
+export * from "@/tables/methods/createTableMethods";
+export * from "@/tables/methods/createTableWatcher";
+
 export * from "@/tables/types";
